requirements.txt
Flask==2.0.3
scikit-learn==1.0.2
spacy==3.4.1
sentence-transformers==2.2.2
textblob==0.17.1
numpy==1.22.4
python-dotenv==0.19.0